+++ examples/LDL_weak_1level.cpp
@@ -26,14 +26,15 @@ class BLR2_SPD {
 public:
   Hatrix::RowColMap<Hatrix::Matrix> D, S;
   Hatrix::RowMap U, Uc;
-  int N, block_size, n_blocks, rank, admis;
+  int64_t N, block_size, n_blocks, rank, admis;
   double construct_error;
 
-  BLR2_SPD(const randvec_t& randpts, int N, int block_size, int rank, int admis):
+  BLR2_SPD(const randvec_t& randpts, int64_t N, int64_t block_size, int64_t rank,
+           int64_t admis):
     N(N), block_size(block_size), n_blocks(N/block_size), rank(rank), admis(admis)
   {
-    for (int i = 0; i < n_blocks; ++i) {
-      for (int j = 0; j < n_blocks; ++j) {      
+    for (int64_t i = 0; i < n_blocks; ++i) {
+      for (int64_t j = 0; j < n_blocks; ++j) {
 	D.insert(i, j,
 		   Hatrix::generate_laplacend_matrix(randpts,
 						     block_size, block_size,
@@ -91,7 +92,7 @@ public:
     Uf_splits[1] = U[row];
     return Uf;
   }
-  
+
 };
 
 Hatrix::Matrix copy_dense(BLR2_SPD& A) {
@@ -133,7 +134,7 @@ void partial_ldl_diag(BLR2_SPD& A, int i) {
 
 Hatrix::Matrix factorize(BLR2_SPD& A) {
   assert(A.rank < A.block_size);
-  
+
   for(int64_t i = 0; i < A.n_blocks; i++) {
     A.Uc.insert(i, make_complement(A.U[i]));
     //Transform diagonal
@@ -165,7 +166,7 @@ Hatrix::Matrix factorize(BLR2_SPD& A) {
 
 void substitute(BLR2_SPD& A, Hatrix::Matrix& root, Hatrix::Matrix& b) {
   assert(A.rank < A.block_size);
-  
+
   //Split b
   int c_size = A.block_size - A.rank;
   vec local_split_indices{c_size};
